Implement authorization policies for Attendee and Event models; update authorization gates in controllers

+++ event-management/app/Http/Controllers/Api/EventController.php
@@ -19,7 +19,7 @@ class EventController extends Controller
      */
     public function index()
     {
-        // Gate::authorize('viewAny', Event::class);
+        Gate::authorize('viewAny', Event::class);
         //Querybuilder starten
         $query = $this->loadRelationships(Event::query());
         // Relationen definieren, die möglich sein sollen
@@ -38,7 +38,7 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
-        // Gate::authorize('create', Event::class);
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -57,7 +57,7 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-        // Gate::authorize('view', $event);
+        Gate::authorize('view', $event);
         return new EventResource($this->loadRelationships($event));
     }
 
@@ -66,7 +66,7 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
-        Gate::authorize('update-event', $event);
+        Gate::authorize('update', $event);
         $event->update([
             // sometimes => Bedeutet, dass der Rest der Validierungen danach nur geprüft wird,
             // wenn das Wertepaar auch mit übergeben wird.
@@ -85,7 +85,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
-        // Gate::authorize('delete', $event);
+        Gate::authorize('delete', $event);
         $event->delete();
 
         return response()->json([
